% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helpers.R
\name{diplay_sql_with_comments}
\alias{diplay_sql_with_comments}
\title{Display sql with comments}
\usage{
diplay_sql_with_comments(file)
}
\arguments{
\item{file}{The name of the sql file you want to display.}
}
\description{
Display sql with comments
}
