% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helpers.R
\name{display_sql_no_comments}
\alias{display_sql_no_comments}
\title{Display sql code with comments removed}
\usage{
display_sql_no_comments(file)
}
\arguments{
\item{file}{The name of the sql file you want to display.}
}
\description{
\code{display_sql_no_comments()} will display a sql file with the comments removed.
Any comments indicated by the double dash -- will be removed. Multiline comments
will not be removed.
}
