% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conn.R
\name{get_conn}
\alias{get_conn}
\title{Title}
\usage{
get_conn(dsn)
}
\arguments{
\item{dsn}{A DSN entry: edify, PROD, REPT, BRPT, etc...}
}
\value{
a database connection
}
\description{
Title
}
\examples{
get_conn('edify')
get_conn('REPT')
}
