% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_data_from_pin}
\alias{get_data_from_pin}
\title{Get data from a pin stored in RStudio Connect}
\usage{
get_data_from_pin(pin_name)
}
\arguments{
\item{pin_name}{The name of the pin, as stored on RStudio Connect instance.}
}
\value{
a dataframe containing all data from requested pin
}
\description{
Get data from a pin stored in RStudio Connect
}
