% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_data_from_pin2}
\alias{get_data_from_pin2}
\title{Get data from a pin stored in the new Posit Connect Server}
\usage{
get_data_from_pin2(pin_name)
}
\arguments{
\item{pin_name}{The name of the pin, as stored on Posit Connect instance.}
}
\value{
a dataframe containing all data from requested pin
}
\description{
Get data from a pin stored in the new Posit Connect Server
}
