% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_get_data_from_pin.R
\name{get_data_from_pin_local}
\alias{get_data_from_pin_local}
\title{get_data_from_pin_local
A helper function for get_data_from_pins()}
\usage{
get_data_from_pin_local(pin_name, prod = TRUE)
}
\arguments{
\item{pin_name}{a pin already hosted on one of the Posit Servers}

\item{prod}{defaulted to true, is for the user to pick which server to pull pins from locally. If false it will pull pins from the test server}
}
\value{
data stored as a pin
}
\description{
get_data_from_pin_local
A helper function for get_data_from_pins()
}
