% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_get_data_from_pins.R
\name{get_data_from_pin_prod}
\alias{get_data_from_pin_prod}
\title{get_data_from_pin_prod
A helper function for get_data_from_pins()}
\usage{
get_data_from_pin_prod(pin_name)
}
\arguments{
\item{pin_name}{a pin already hosted on one of the Posit Servers}
}
\value{
data stored as a pin
}
\description{
get_data_from_pin_prod
A helper function for get_data_from_pins()
}
