% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_get_data_from_pin.R
\name{get_data_from_pin_test}
\alias{get_data_from_pin_test}
\title{get_data_from_pin_test
A helper function for get_data_from_pins()}
\usage{
get_data_from_pin_test(pin_name)
}
\arguments{
\item{pin_name}{a pin already hosted on one of the Posit Servers}
}
\value{
data stored as a pin
}
\description{
get_data_from_pin_test
A helper function for get_data_from_pins()
}
