% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_data_from_sql_url}
\alias{get_data_from_sql_url}
\title{Get data from a sql url}
\usage{
get_data_from_sql_url(query_url, dsn)
}
\arguments{
\item{query_url}{the url of a sql query}

\item{dsn}{A DSN entry: edify, PROD, REPT, BRPT, etc...}
}
\value{
a dataframe containing the results of your sql query
}
\description{
Get data from a sql url
}
