% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_objects.R
\name{get_db_connection_object}
\alias{get_db_connection_object}
\title{Get Database Connection Object}
\usage{
get_db_connection_object(dsn)
}
\arguments{
\item{dsn}{A string of text representing the name of a DSN entry}
}
\value{
either a DBMS connection object, or an ODBC connection object, from DBI::dbConnect
}
\description{
Get Database Connection Object
}
