% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_objects.R
\name{get_pins_connection_object}
\alias{get_pins_connection_object}
\title{Get Pins Connection Object}
\usage{
get_pins_connection_object()
}
\value{
A connection object to read and write pins to on RStudio Connect
}
\description{
Get Pins Connection Object
}
