% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_data_from_csv}
\alias{load_data_from_csv}
\title{Load data from a CSV formatted file}
\usage{
load_data_from_csv(file_name)
}
\arguments{
\item{file_name}{the name of the file, assumed to be located in a directory called 'data'}
}
\value{
A dataframe with all content of the CSV file.
}
\description{
Load data from a CSV formatted file
}
