% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_data_from_rds}
\alias{load_data_from_rds}
\title{Get data from a rds source.}
\usage{
load_data_from_rds(file_name)
}
\arguments{
\item{file_name}{The rds file you want to load}
}
\value{
a dataframe containing the data from the .rds file
}
\description{
Get data from a rds source.
}
