% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_data_from_xlsx}
\alias{load_data_from_xlsx}
\title{Load data from an Excel spreadsheet formatted file}
\usage{
load_data_from_xlsx(file_name)
}
\arguments{
\item{file_name}{the name of the file, assumed to be located in a directory called 'data'}
}
\value{
A dataframe with all content of the Excel spreadsheet.
}
\description{
Load data from an Excel spreadsheet formatted file
}
