% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mung_dataframe.R
\name{mung_dataframe}
\alias{mung_dataframe}
\title{Munge dataframe}
\usage{
mung_dataframe(df)
}
\arguments{
\item{df}{}
}
\value{
a clean dataframe
}
\description{
Munge dataframe
}
\examples{
munge_dataframe(bob)
}
