% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{save_data_as_file}
\alias{save_data_as_file}
\title{Save Data as File}
\usage{
save_data_as_file(
  input_df,
  file_name,
  delim = "|",
  with_header = FALSE,
  quote = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Note: can contain any arbitrary data fields.}

\item{file_name}{The name of the file that will be saved (e.g. "my_data.txt").}

\item{delim}{A character that will be placed between data fields, in the output file (e.g. "|").}

\item{with_header}{Boolean. Will determine if the column names of input_df will be included in the output file.}

\item{quote}{Boolean. If TRUE will wrap each data column in double quotes.}
}
\value{
Returns nothing.
}
\description{
Save Data as File
}
\details{
** Saves data to local machine. **

The default location for the file to be saved is in a folder called "sensitive".
If this folder does not exist, this function will not work.
}
