% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{save_data_as_pin}
\alias{save_data_as_pin}
\title{Save Data as Pin}
\usage{
save_data_as_pin(input_df, pin_name)
}
\arguments{
\item{input_df}{A Data Frame. Note: can contain any arbitrary data fields.}

\item{pin_name}{The name of the pin to be saved.}
}
\value{
Returns nothing.
}
\description{
Save Data as Pin
}
