% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_banner_credentials.R
\name{set_banner_db_password}
\alias{set_banner_db_password}
\title{Set Banner DB Password}
\usage{
set_banner_db_password()
}
\description{
\code{set_edify_password} will prompt you with a dialog box to input your new password
for Banner DB.
}
\details{
\strong{To reset your Banner DB password.}
}
