% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_banner_credentials.R
\name{set_banner_db_username}
\alias{set_banner_db_username}
\title{Set Banner DB Username}
\usage{
set_banner_db_username()
}
\description{
\code{set_banner_db_username} will prompt you with a dialog box to input your user name
for Banner DB.
}
\details{
**To set your Banner DB user name. **
}
