% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edify_credentials.R
\name{set_edify_password}
\alias{set_edify_password}
\title{Set Edify Password}
\usage{
set_edify_password()
}
\description{
\code{set_edify_password} will prompt you with a dialog box to input your new password
for Edify DB connections.
}
\details{
\strong{To reset your Edify password.}

\itemize{
\item Navigate to \href{https://utahtech.edh.eab.com/}{Edify}.
\item Select the SSO login.
\item Login with your UT single sign-on credentials.
\item Click the question mark in the upper-right corner.
\item Select User Profile.
\item Select Generate new credential.
\item New password can be found here.
}
}
