% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_pins_api_key}
\alias{set_pins_api_key}
\title{Set API key for pins}
\usage{
set_pins_api_key()
}
\description{
\code{set_pins_api_key} will prompt you with a dialog box to input the API key for pins access.
}
\details{
**To set the API key to access pins. **
}
