% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_sis_db_password}
\alias{set_sis_db_password}
\title{Set SIS DB Password}
\usage{
set_sis_db_password()
}
\description{
\code{set_sis_db_password} will prompt you with a dialog box to input your new password
for SIS DB.
}
\details{
\strong{To reset your SIS DB password.}
}
