% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sis_db_credentials.R
\name{set_sis_db_username}
\alias{set_sis_db_username}
\title{Set SIS DB User Name}
\usage{
set_sis_db_username()
}
\description{
\code{set_sis_db_username} will prompt you with a dialog box to input your user name
for SIS DB.
}
\details{
**To set your SIS DB user name. **
}
