% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmgr_funs.R
\name{uth_check_pkg_mgr_status}
\alias{uth_check_pkg_mgr_status}
\title{Utah Tech check package manager status}
\usage{
uth_check_pkg_mgr_status()
}
\value{
prints the current package manager URL and R version.
}
\description{
The \code{uth_check_pkg_mgr_status} function retrieves the list of available packages, converts it into a data frame, and selects the first row. It then extracts the Repository column value from this row and prints the current package manager URL. This function is useful for checking the URL of the package repository currently being used.
}
