% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_helpers.R,
%   R/uth_clean_student_id.R
\name{uth_clean_student_id}
\alias{uth_clean_student_id}
\title{Clean Student ID}
\usage{
uth_clean_student_id(student_id)

uth_clean_student_id(student_id)
}
\arguments{
\item{student_id}{A student id as a number, email, text with a leading d or D.}
}
\value{
A character vector of cleaned student IDs, each formatted to be 8 characters long, with leading zeros if necessary. The function removes any leading "D" or "d" and handles email addresses by truncating them to the first 8 characters.

A dataframe with student ids to manipulated into a character format without a leading "D", This can also check if a student id is a slate_id or a banner_id if it has leading alpha from nsc house tables
}
\description{
This function processes student IDs to ensure they are in a consistent character format without leading "D" or "d". It can handle various formats of student IDs, including numeric IDs, email addresses, and IDs with leading characters.
}
\details{
This function standardizes student IDs by:
\itemize{
\item Removing leading "D" or "d" characters.
\item Handling email addresses by extracting the first 8 characters.
\item Converting numeric IDs to character format.
\item Padding IDs with leading zeros to ensure they are 8 characters long.
}

This function is useful for preparing student IDs for analysis or integration with other systems that require a standardized format.
}
\examples{
\dontrun{
# Example usage:
ids <- c("D1234567", "d2345678", "BO345678", "D1234567@utahtech.edu", 123456)
cleaned_ids <- uth_clean_student_id(ids)

# Using the function within a dplyr pipeline
df <- data.frame(student_id = ids)
df <- df \%>\%
  dplyr::mutate(cleaned_student_id = uth_clean_student_id(student_id))
}

}
