% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmgr_funs.R
\name{uth_download_r_for_mac}
\alias{uth_download_r_for_mac}
\title{Utah Tech download R for Mac OS}
\usage{
uth_download_r_for_mac(suggested_version, mac_type)
}
\arguments{
\item{suggested_version}{user input of a desired version of R example "R.4.4.1" as a string}

\item{mac_type}{a string signifying user's Mac processor type, currently supports "M1" or "Intel"}
}
\value{
a download R file from CRAN in the user's download folder.
}
\description{
The \code{uth_download_r_for_mac} function constructs a download URL for the specified version of R based on the type of Mac (either "M1" or other). It sets the base URL for CRAN, appends the appropriate architecture specification (-arm64 for M1 or -x86_64 for others), and forms the complete download link and file name. The function then determines the download folder path, constructs the destination file path, and attempts to download the file using download.file, handling any errors with tryCatch. If successful, it informs the user of the download location; otherwise, it provides an error message. This function simplifies downloading the correct version of R for different Mac architectures.
}
