% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_get_data_from_pin.R
\name{uth_get_data_from_pin}
\alias{uth_get_data_from_pin}
\title{get_data_from_pin()
This R function retrieves pins data from a Posit server. It uses the pins package to connect to the server, authenticate, and fetch the required data. This function ensures secure and efficient data retrieval from a Posit server, making it easy to integrate pins data into your R workflows.}
\usage{
uth_get_data_from_pin(pin_name, prod = TRUE)
}
\arguments{
\item{pin_name}{The name of the pin to retrieve hosted on one of Utah Tech's Posit Servers}

\item{prod}{defaulted to true to pull pins from the prod Posit Server, if false it will pull pins from the test server. This parameter is only used when retrieving pins locally.}
}
\value{
The name of the pin to retrieve
}
\description{
get_data_from_pin()
This R function retrieves pins data from a Posit server. It uses the pins package to connect to the server, authenticate, and fetch the required data. This function ensures secure and efficient data retrieval from a Posit server, making it easy to integrate pins data into your R workflows.
}
