% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmgr_funs.R
\name{uth_install_from_pkg_mgr}
\alias{uth_install_from_pkg_mgr}
\title{Utah Tech install from Posit package manager}
\usage{
uth_install_from_pkg_mgr(update_and_install = FALSE)
}
\arguments{
\item{update_and_install}{It has a single parameter update_and_install which defaults to FALS which will only update currently installed packages. If you want to install new packages use please use TRUE.}
}
\value{
Loads the saved package information.
}
\description{
The uth_install_from_pkg_mgr function manages and updates R packages based on previously saved package information. It first fetches the list of available packages and loads the saved package data. It then joins this data with the available packages to identify which ones are installed. After deleting the temporary file containing the saved package information, the function updates the installed packages. If the update_and_install parameter is set to TRUE, it also re-installs the packages. This ensures that your R environment is up-to-date with the latest versions of your installed packages, providing an efficient way to manage package updates and installations.
}
