% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmgr_funs.R
\name{uth_load_my_pkg_info}
\alias{uth_load_my_pkg_info}
\title{Utah Tech load my package information}
\usage{
uth_load_my_pkg_info()
}
\value{
This function complements the save_my_pkg_info function by providing a way to retrieve the saved package information.
}
\description{
The \code{uth_load_my_pkg_info} function first gets the path to the temporary directory and defines the file path for my_packages_info.rda. It checks if this file exists; if it does, the function loads the file, displays a success message, and returns the my_packages_info data. If the file does not exist, it issues a warning message indicating that the save_my_packages_info function should be run first and returns NULL. This function is designed to load previously saved package information from a temporary file.
}
