% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_make_outcome_average.R
\name{uth_make_outcome_average}
\alias{uth_make_outcome_average}
\title{Aggregate Column Averages in a Data Frame}
\usage{
uth_make_outcome_average(.data, avg_var, ..., dp = 2)
}
\arguments{
\item{.data}{A data frame containing the data to be summarized.}

\item{avg_var}{A string representing the name of the variable for which the summary statistics will be calculated.}

\item{...}{One or more grouping variables by which the data will be grouped before calculating the summary statistics.}

\item{dp}{An integer specifying the number of decimal places to which the summary statistics should be rounded. The default value is 2.}
}
\value{
A data frame with the aggregated averagegs of the specified column, grouped by the additional columns.
}
\description{
Calculates summary statistics (mean, standard deviation, and median) for a specified variable within a data frame, grouped by one or more additional variables. The results are rounded to a specified number of decimal places.
}
