% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmgr_funs.R
\name{uth_save_my_pkg_info}
\alias{uth_save_my_pkg_info}
\title{Utah Tech save my packages information}
\usage{
uth_save_my_pkg_info()
}
\value{
the file path of the temporary directory saved data frame.
}
\description{
The \code{uth_save_my_pkg_info} function creates a data frame of installed packages, excluding base packages, and adds a column is_installed set to TRUE. It then gets the path to the temporary directory and defines the file path for my_packages_info.rda. The function saves the data frame to this file and informs the user that the file has been successfully saved, returning the file path for later retrieval. This function is designed to save the current package information to a temporary file for future use.
}
