% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_update_utpackages.R
\name{uth_update_utpackages}
\alias{uth_update_utpackages}
\title{uth_update_utpackages}
\usage{
uth_update_utpackages(utpackage)
}
\arguments{
\item{utpackage}{A package on made by Utah Tech Univeristy housed on a Github repo}
}
\value{
A package made by Utah Tech Univeristy housed on a Github
}
\description{
uth_update_utpackages
}
