% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conn.R
\name{get_conn}
\alias{get_conn}
\title{Set up a database connection}
\usage{
get_conn(dsn)
}
\arguments{
\item{dsn}{A DSN entry: edify, PROD, REPT, BRPT, etc...}
}
\value{
a database connection
}
\description{
Set up a database connection
}
\examples{

}
