% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frome_sql_file.R
\name{get_data_from_sql_file}
\alias{get_data_from_sql_file}
\title{Get data from a sql file}
\usage{
get_data_from_sql_file(file_name, dsn)
}
\arguments{
\item{file_name}{the name of the sql file that you want to use}

\item{dsn}{A DSN entry: edify, PROD, REPT, BRPT, etc...}
}
\value{
a dataframe containing the results of your sql query
}
\description{
Get data from a sql file
}
