% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_util.R
\name{make_standard_folders}
\alias{make_standard_folders}
\title{Make a directory structure for a standard project}
\usage{
make_standard_folders()
}
\description{
\code{make_standard_folders()} will create our standard directory structure in the
file you are in when you run it. You should run this function when you are
setting up a project. When you run \code{make_standard_folders()} you should be in
the same directory as your .Rproj file.
}
