% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edify_password.R
\name{set_edify_password}
\alias{set_edify_password}
\title{Set Edify password}
\usage{
set_edify_password()
}
\description{
\code{set_edify_password} will prompt you with a dialog box to input your new password
from edify.
}
\details{
\strong{To reset your Edify password}

\itemize{
\item Navigate to \href{https://dixie.edh.eab.com/}{Edify}.
\item Select the SSO login.
\item Login with your UT single sign-on credentials.
\item Click the question mark in the upper left corner.
\item Select User Profile
\item Select Generate new credential.
}
}
