% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edify_credentials.R
\name{set_edify_username}
\alias{set_edify_username}
\title{Set Edify username}
\usage{
set_edify_username()
}
\description{
\code{set_edify_username} will prompt you with a dialog box to input your username
from Edify.
}
\details{
\strong{To set your Edify username}

\itemize{
\item Navigate to \href{https://utahtech.db.edh.eab.com/}{Edify}.
\item Select the SSO login.
\item Login with your UT single sign-on credentials.
\item Click the question mark in the upper-right corner.
\item Select User Profile.
\item Edify username can be found here.
}
}
