% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_clean_student_id.R
\name{uth_clean_student_id}
\alias{uth_clean_student_id}
\title{uth_clean_student_id}
\usage{
uth_clean_student_id(student_id)
}
\arguments{
\item{student_id}{A student id as a number, email, text with a leading d or D.}
}
\value{
A dataframe with student ids to manipulated into a character format without a leading "D", This can also check if a student id is a slate_id or a banner_id if it has leading alpha from nsc house tables
}
\description{
uth_clean_student_id
}
