% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uth_make_outcome_count.R
\name{uth_make_outcome_count}
\alias{uth_make_outcome_count}
\title{uth_make_outcome_count}
\usage{
uth_make_outcome_count(.data, ...)
}
\arguments{
\item{.data}{A data frame with columns you would like to aggregate on}

\item{...}{the columns you would lke to aggregate over}
}
\description{
uth_make_outcome_count
}
